[refactor]: getSubscribedClubs를 ClubQueryUseCase로 이동
