+++ src/main/java/com/kustacks/kuring/club/application/port/in/ClubQueryUseCase.java
@@ -1,18 +1,15 @@
 package com.kustacks.kuring.club.application.port.in;
 
-import com.kustacks.kuring.club.application.port.in.dto.ClubDetailCommand;
-import com.kustacks.kuring.club.application.port.in.dto.ClubDetailResult;
-import com.kustacks.kuring.club.application.port.in.dto.ClubDivisionResult;
-import com.kustacks.kuring.club.application.port.in.dto.ClubListCommand;
-import com.kustacks.kuring.club.application.port.in.dto.ClubListResult;
+import com.kustacks.kuring.club.application.port.in.dto.*;
 
-import java.util.List;
+import java.util.*;
 
 public interface ClubQueryUseCase {
     List<ClubDivisionResult> getClubDivisions();
 
     ClubListResult getClubs(ClubListCommand command);
 
-    ClubDetailResult getClubDetail(ClubDetailCommand command);
+    ClubListResult getSubscribedClubs(SubscribedClubListCommand command);
 
+    ClubDetailResult getClubDetail(ClubDetailCommand command);
 }
