+++ src/main/java/com/kustacks/kuring/club/application/port/in/ClubSubscriptionUseCase.java
@@ -1,14 +1,10 @@
 package com.kustacks.kuring.club.application.port.in;
 
-import com.kustacks.kuring.club.application.port.in.dto.ClubListResult;
-import com.kustacks.kuring.club.application.port.in.dto.ClubSubscriptionCommand;
-import com.kustacks.kuring.club.application.port.in.dto.SubscribedClubListCommand;
+import com.kustacks.kuring.club.application.port.in.dto.*;
 
 public interface ClubSubscriptionUseCase {
 
     long addSubscription(ClubSubscriptionCommand command);
 
     long removeSubscription(ClubSubscriptionCommand command);
-
-    ClubListResult getSubscribedClubs(SubscribedClubListCommand command);
 }
